Leverage gproc and genserver
* Drop use of GenEvent in Music Service. Instead I opt for a user by
  user genServer which functions as a proxy for a future live music
  service connection.
* Update calls in EavesdropOTP.Worker to call to new MusicService API.
* Put main body of the service under a simple_one_to_one supervisor so
  that each user gets their own instance of an FSM to represent their
  listening state and a mock music service

TODO: Ensure that the FSM still works. I am initializing in the idle
state, because by starting an eavesdrop child I am effectively "logging
in". There is no such thing as a live eavesdrop process for which there
does not exist a logged in user. As such, I need to cleanup processes
once a user logs out and ensure that they are de-referenced in gproc