convert to new gen_statem module
I think the docs are wrong regarding the return tuple for the init
function. Docs say it needs to be of the form (erlang) {ok, State,
Data}, but what I am observing is that the first value needs to be an
atom corresponding to the callback_module atom. This require further
investigation and potentially a doc fix